libipt: Add a new cr3 callback
This patch adds the capability to libipt to load executable files lazily. This is
useful for a PT implementation that enumerates all processes first, but
may not encounter all enumerated processes in the actual trace.
Currently all these binaries would need to be loaded first. Allow
to do it lazily by adding a callback when a new CR3 is encountered.
This way we can save a lot of start up time and potential errors.

Signed-off-by: Andi Kleen <ak@linux.intel.com>